requirements.txt
torch>=2.0
torchvision
gymnasium>=0.29.1
stable-baselines3[extra]>=2.3.0
pygame>=2.5.0
numpy>=1.24.0
tensorboard>=2.14.0
matplotlib
pyyaml
